Adds comprehensive error handling test coverage
Adds test coverage for critical error scenarios including missing PDF text, translation failures, network issues fetching models, and missing synthesized files.
Ensures proper error logging and graceful handling of edge cases in production.
Improves robustness and reliability of the text-to-speech synthesis pipeline.
